requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.4.26
click==7.1.2
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==1.1.4
Flask-APScheduler==1.12.4
flask-mongoengine==1.0.0
Flask-Pydantic==0.13.1
Flask-WTF==1.2.2
greenlet==3.2.2
idna==3.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.3.10
MarkupSafe==2.0.1
mongoengine==0.29.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==1.0.1
WTForms==3.2.1
